💄  Style: 검색결과 페이지 컨테이너 디자인 개선
